Consolidate agent protocol utilities
